feat : 월근태현황 조회 기능 수정 완료  / 2023-09-25 이세은

+++ KccBrew/src/main/java/kr/co/kccbrew/schdlMng/model/AsResultVo.java
@@ -14,7 +14,7 @@ public class AsResultVo {
 	private Integer storeManagerFeedBack;
 	private Date resultDate;
 	private Integer fileSeq;
-	private Integer asPrice;
+	private double asPrice;
 	private String modUser;
 	private Date modDate;
 	private String reApply;
